# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/unidade": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Lista informaçãoes de todas as Unidades",
                "description": "Exibe todas as Unidades com suas Lotação",
                "operationId": "2250b14b758c20a6048832f40e52f176",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº de páginas",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Cadastra uma nova Unidade",
                "description": "Registro de uma nova Unidade",
                "operationId": "0bbdbda9ace1e68eb6d131c76fe08b79",
                "parameters": [
                    {
                        "name": "unid_nome",
                        "in": "query",
                        "description": "Nome da Unidade",
                        "required": true
                    },
                    {
                        "name": "unid_sigla",
                        "in": "query",
                        "description": "Sigla da Unidade",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unidade/{unid_id}": {
            "get": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Mostra informações de uma Unidade específica",
                "description": "Pesquisa por uma Unidade através do (unid_id)",
                "operationId": "bacd89162fdfcece3328dc531df429e1",
                "parameters": [
                    {
                        "name": "unid_id",
                        "in": "path",
                        "description": "Nº de identificação da Unidae",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Atualizar dados de uma Unidade",
                "description": "Editar os dados de uma Unidade através do (unid_id)",
                "operationId": "5f2bd24c793d11dd8e43eec35a810be3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "unid_id",
                                    "unid_nome",
                                    "unid_sigla"
                                ],
                                "properties": {
                                    "unid_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "unid_nome": {
                                        "type": "string",
                                        "example": "Pirinópolis"
                                    },
                                    "unid_sigla": {
                                        "type": "string",
                                        "example": "SIGLA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unidade atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unidade atualizada com sucesso"
                                        },
                                        "unidade": {
                                            "properties": {
                                                "unid_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "unid_nome": {
                                                    "type": "string",
                                                    "example": "Pirinópolis"
                                                },
                                                "unid_sigla": {
                                                    "type": "string",
                                                    "example": "MT"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unidades"
                ],
                "summary": "Exclui uma Unidade",
                "description": "Exclui uma Unidade através do (unid_id)",
                "operationId": "0e13d61a5bb0490f5ebe46beb16c3654",
                "parameters": [
                    {
                        "name": "unid_id",
                        "in": "path",
                        "description": "ID da Unidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unidade excluída com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir a Unidade"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}